% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis.R
\name{thesis}
\alias{thesis}
\title{Rmarkdown Thesis Template at Muenster University of Applied Sciences}
\usage{
thesis(..., keep_tex = TRUE)
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document}}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}
}
\value{
R Markdown output format to pass to
  \code{\link[rmarkdown:render]{render}}
}
\description{
Create your thesis with all the beauty of R.
}
\examples{

\dontrun{
library(rmarkdown)
draft("MyThesis.Rmd", template = "muas_thesis", package = "muasdown")
}

}
